fixing page search result
